refactor(DEV-001): Remove db.schema.ts and update imports to use individual schema files
- Updated config/auth.ts to import from individual schema files
- Updated controllers/auth.controller.ts to import from schemas/user.schema.ts
- Updated websockets/auth.middleware.ts to import from schemas/user.schema.ts and fixed dbClient shadowing
- Updated infrastructure/better-auth.client.ts to import from individual schema files
- Deleted infrastructure/db.schema.ts (re-export file no longer needed)
- Changed import path in better-auth.client.ts from './db.schema' to '../schemas/'

This change follows the constraint: 'no index.ts files in folders' and makes
dependencies more explicit and traceable. Type-check errors reduced from 77 to 18,
with remaining 18 errors in incomplete service stub implementations (BE-010, BE-011, BE-012).
